+++ bonobo/config/__init__.py
@@ -1,13 +1,16 @@
 from bonobo.config.configurables import Configurable
-from bonobo.config.options import Option, Method
+from bonobo.config.options import Method, Option
 from bonobo.config.processors import ContextProcessor
-from bonobo.config.services import Container, Service
+from bonobo.config.services import Container, Exclusive, Service, requires
 
+# bonobo.config public programming interface
 __all__ = [
     'Configurable',
     'Container',
     'ContextProcessor',
-    'Option',
+    'Exclusive',
     'Method',
+    'Option',
     'Service',
+    'requires',
 ]