+++ bonobo/examples/tutorials/tut01e02.py
@@ -1,0 +1,14 @@
+import bonobo
+
+graph = bonobo.Graph(
+    [
+        'foo',
+        'bar',
+        'baz',
+    ],
+    str.upper,
+    print,
+)
+
+if __name__ == '__main__':
+    bonobo.run(graph)