+++ bonobo/nodes/basics.py
@@ -1,9 +1,12 @@
 import functools
 from pprint import pprint as _pprint
 
+import itertools
 from colorama import Fore, Style
 
-from bonobo.config import Configurable, ContextProcessor, Option
+from bonobo import settings
+from bonobo.config import Configurable, Option
+from bonobo.config.processors import ContextProcessor
 from bonobo.structs.bags import Bag
 from bonobo.util.objects import ValueHolder
 from bonobo.util.term import CLEAR_EOL
@@ -15,7 +18,7 @@ __all__ = [
     'Tee',
     'count',
     'pprint',
-    'PrettyPrint',
+    'PrettyPrinter',
     'noop',
 ]
 
@@ -68,7 +71,24 @@ def _count_counter(self, context):
     context.send(Bag(counter._value))
 
 
-pprint = Tee(_pprint)
+class PrettyPrinter(Configurable):
+    def call(self, *args, **kwargs):
+        formater = self._format_quiet if settings.QUIET else self._format_console
+
+        for i, (item, value) in enumerate(itertools.chain(enumerate(args), kwargs.items())):
+            print(formater(i, item, value))
+
+    def _format_quiet(self, i, item, value):
+        return ' '.join(((' ' if i else '-'), str(item), ':', str(value).strip()))
+
+    def _format_console(self, i, item, value):
+        return ' '.join(
+            ((' ' if i else '•'), str(item), '=', str(value).strip().replace('\n', '\n' + CLEAR_EOL), CLEAR_EOL)
+        )
+
+
+pprint = PrettyPrinter()
+pprint.__name__ = 'pprint'
 
 
 def PrettyPrint(title_keys=('title', 'name', 'id'), print_values=True, sort=True):