+++ bonobo/nodes/io/csv.py
@@ -3,8 +3,9 @@ import csv
 from bonobo.config import Option
 from bonobo.config.processors import ContextProcessor
 from bonobo.constants import NOT_MODIFIED
+from bonobo.nodes.io.file import FileReader, FileWriter
+from bonobo.nodes.io.base import FileHandler, IOFormatEnabled
 from bonobo.util.objects import ValueHolder
-from .file import FileHandler, FileReader, FileWriter
 
 
 class CsvHandler(FileHandler):
@@ -28,7 +29,7 @@ class CsvHandler(FileHandler):
     headers = Option(tuple)
 
 
-class CsvReader(CsvHandler, FileReader):
+class CsvReader(IOFormatEnabled, FileReader, CsvHandler):
     """
     Reads a CSV and yield the values as dicts.
 
@@ -49,6 +50,7 @@ class CsvReader(CsvHandler, FileReader):
 
         if not headers.get():
             headers.set(next(reader))
+        _headers = headers.get()
 
         field_count = len(headers)
 
@@ -60,17 +62,18 @@ class CsvReader(CsvHandler, FileReader):
             if len(row) != field_count:
                 raise ValueError('Got a line with %d fields, expecting %d.' % (len(row), field_count, ))
 
-            yield dict(zip(headers.value, row))
+            yield self.get_output(dict(zip(_headers, row)))
 
 
-class CsvWriter(CsvHandler, FileWriter):
+class CsvWriter(IOFormatEnabled, FileWriter, CsvHandler):
     @ContextProcessor
     def writer(self, context, fs, file, lineno):
         writer = csv.writer(file, delimiter=self.delimiter, quotechar=self.quotechar, lineterminator=self.eol)
         headers = ValueHolder(list(self.headers) if self.headers else None)
         yield writer, headers
 
-    def write(self, fs, file, lineno, writer, headers, row):
+    def write(self, fs, file, lineno, writer, headers, *args, **kwargs):
+        row = self.get_input(*args, **kwargs)
         if not lineno:
             headers.set(headers.value or row.keys())
             writer.writerow(headers.get())