+++ bonobo/nodes/io/file.py
@@ -1,63 +1,11 @@
-from bonobo.config import Option, Service
-from bonobo.config.configurables import Configurable
+from bonobo.config import Option
 from bonobo.config.processors import ContextProcessor
 from bonobo.constants import NOT_MODIFIED
+from bonobo.nodes.io.base import FileHandler, Reader, Writer
 from bonobo.util.objects import ValueHolder
 
-__all__ = [
-    'FileReader',
-    'FileWriter',
-]
 
-
-class FileHandler(Configurable):
-    """Abstract component factory for file-related components.
-    
-    Args:
-        path (str): which path to use within the provided filesystem.
-        eol (str): which character to use to separate lines.
-        mode (str): which mode to use when opening the file.
-        fs (str): service name to use for filesystem.
-    """
-
-    path = Option(str, required=True, positional=True)  # type: str
-    eol = Option(str, default='\n')  # type: str
-    mode = Option(str)  # type: str
-
-    fs = Service('fs')  # type: str
-
-    @ContextProcessor
-    def file(self, context, fs):
-        with self.open(fs) as file:
-            yield file
-
-    def open(self, fs):
-        return fs.open(self.path, self.mode)
-
-
-class Reader(FileHandler):
-    """Abstract component factory for readers.
-    """
-
-    def __call__(self, *args):
-        yield from self.read(*args)
-
-    def read(self, *args):
-        raise NotImplementedError('Abstract.')
-
-
-class Writer(FileHandler):
-    """Abstract component factory for writers.
-    """
-
-    def __call__(self, *args):
-        return self.write(*args)
-
-    def write(self, *args):
-        raise NotImplementedError('Abstract.')
-
-
-class FileReader(Reader):
+class FileReader(Reader, FileHandler):
     """Component factory for file-like readers.
 
     On its own, it can be used to read a file and yield one row per line, trimming the "eol" character at the end if
@@ -75,7 +23,7 @@ class FileReader(Reader):
             yield line.rstrip(self.eol)
 
 
-class FileWriter(Writer):
+class FileWriter(Writer, FileHandler):
     """Component factory for file or file-like writers.
 
     On its own, it can be used to write in a file one line per row that comes into this component. Extending it is
@@ -89,11 +37,11 @@ class FileWriter(Writer):
         lineno = ValueHolder(0)
         yield lineno
 
-    def write(self, fs, file, lineno, row):
+    def write(self, fs, file, lineno, line):
         """
         Write a row on the next line of opened file in context.
         """
-        self._write_line(file, (self.eol if lineno.value else '') + row)
+        self._write_line(file, (self.eol if lineno.value else '') + line)
         lineno += 1
         return NOT_MODIFIED