+++ bonobo/nodes/io/json.py
@@ -3,45 +3,39 @@ from itertools import starmap
 
 from bonobo.structs.bags import Bag
 from bonobo.config.processors import ContextProcessor
-from .file import FileWriter, FileReader
-
-__all__ = [
-    'JsonWriter',
-]
+from bonobo.constants import NOT_MODIFIED
+from bonobo.nodes.io.base import FileHandler, IOFormatEnabled
+from bonobo.nodes.io.file import FileReader, FileWriter
 
 
-class JsonHandler():
+class JsonHandler(FileHandler):
     eol = ',\n'
     prefix, suffix = '[', ']'
 
 
-class JsonReader(JsonHandler, FileReader):
+class JsonReader(IOFormatEnabled, FileReader, JsonHandler):
     loader = staticmethod(json.load)
 
     def read(self, fs, file):
         for line in self.loader(file):
-            yield line
+            yield self.get_output(line)
 
 
-class JsonDictReader(JsonReader):
-    """ not api, don't use or expect breakage. """
-
-    def read(self, fs, file):
-        yield from starmap(Bag, self.loader(file).items())
-
-
-class JsonWriter(JsonHandler, FileWriter):
+class JsonWriter(IOFormatEnabled, FileWriter, JsonHandler):
     @ContextProcessor
     def envelope(self, context, fs, file, lineno):
         file.write(self.prefix)
         yield
         file.write(self.suffix)
 
-    def write(self, fs, file, lineno, row):
+    def write(self, fs, file, lineno, *args, **kwargs):
         """
         Write a json row on the next line of file pointed by ctx.file.
 
         :param ctx:
         :param row:
         """
-        return super().write(fs, file, lineno, json.dumps(row))
+        row = self.get_input(*args, **kwargs)
+        self._write_line(file, (self.eol if lineno.value else '') + json.dumps(row))
+        lineno += 1
+        return NOT_MODIFIED