+++ tests/config/test_methods.py
@@ -28,7 +28,10 @@ def test_define_with_decorator():
     def Concrete(self, *args, **kwargs):
         calls.append((args, kwargs, ))
 
+    assert callable(Concrete.handler)
     t = Concrete('foo', bar='baz')
+
+    assert callable(t.handler)
     assert len(calls) == 0
     t()
     assert len(calls) == 1
@@ -41,6 +44,7 @@ def test_define_with_argument():
         calls.append((args, kwargs, ))
 
     t = MethodBasedConfigurable('foo', bar='baz', handler=concrete_handler)
+    assert callable(t.handler)
     assert len(calls) == 0
     t()
     assert len(calls) == 1
@@ -54,6 +58,7 @@ def test_define_with_inheritance():
             calls.append((args, kwargs, ))
 
     t = Inheriting('foo', bar='baz')
+    assert callable(t.handler)
     assert len(calls) == 0
     t()
     assert len(calls) == 1
@@ -69,7 +74,9 @@ def test_inheritance_then_decorate():
     def Concrete(self, *args, **kwargs):
         calls.append((args, kwargs, ))
 
+    assert callable(Concrete.handler)
     t = Concrete('foo', bar='baz')
+    assert callable(t.handler)
     assert len(calls) == 0
     t()
     assert len(calls) == 1