+++ tests/io/test_json.py
@@ -1,42 +1,48 @@
 import pytest
 
-from bonobo import Bag, JsonReader, JsonWriter, open_fs
+from bonobo import Bag, JsonReader, JsonWriter, settings
 from bonobo.constants import BEGIN, END
 from bonobo.execution.node import NodeExecutionContext
-from bonobo.util.testing import CapturingNodeExecutionContext
+from bonobo.util.testing import CapturingNodeExecutionContext, FilesystemTester
+
+json_tester = FilesystemTester('json')
+json_tester.input_data = '''[{"x": "foo"},{"x": "bar"}]'''
 
 
-def test_write_json_to_file(tmpdir):
-    fs, filename = open_fs(tmpdir), 'output.json'
+def test_write_json_arg0(tmpdir):
+    fs, filename, services = json_tester.get_services_for_writer(tmpdir)
 
-    writer = JsonWriter(path=filename)
-    context = NodeExecutionContext(writer, services={'fs': fs})
+    with NodeExecutionContext(JsonWriter(filename, ioformat=settings.IOFORMAT_ARG0), services=services) as context:
+        context.write(BEGIN, Bag({'foo': 'bar'}), END)
+        context.step()
 
-    context.start()
-    context.write(BEGIN, Bag({'foo': 'bar'}), END)
-    context.step()
-    context.stop()
-
-    assert fs.open(filename).read() == '[{"foo": "bar"}]'
-
-    with pytest.raises(AttributeError):
-        getattr(context, 'file')
-
-    with pytest.raises(AttributeError):
-        getattr(context, 'first')
+    with fs.open(filename) as fp:
+        assert fp.read() == '[{"foo": "bar"}]'
 
 
-def test_read_json_from_file(tmpdir):
-    fs, filename = open_fs(tmpdir), 'input.json'
-    fs.open(filename, 'w').write('[{"x": "foo"},{"x": "bar"}]')
-    reader = JsonReader(path=filename)
+@pytest.mark.parametrize('add_kwargs', ({}, {
+    'ioformat': settings.IOFORMAT_KWARGS,
+}, ))
+def test_write_json_kwargs(tmpdir, add_kwargs):
+    fs, filename, services = json_tester.get_services_for_writer(tmpdir)
 
-    context = CapturingNodeExecutionContext(reader, services={'fs': fs})
+    with NodeExecutionContext(JsonWriter(filename, **add_kwargs), services=services) as context:
+        context.write(BEGIN, Bag(**{'foo': 'bar'}), END)
+        context.step()
 
-    context.start()
-    context.write(BEGIN, Bag(), END)
-    context.step()
-    context.stop()
+    with fs.open(filename) as fp:
+        assert fp.read() == '[{"foo": "bar"}]'
+
+
+def test_read_json_arg0(tmpdir):
+    fs, filename, services = json_tester.get_services_for_reader(tmpdir)
+
+    with CapturingNodeExecutionContext(
+        JsonReader(filename, ioformat=settings.IOFORMAT_ARG0),
+        services=services,
+    ) as context:
+        context.write(BEGIN, Bag(), END)
+        context.step()
 
     assert len(context.send.mock_calls) == 2